Introduce the `-r' command-line argument
We can now specify a list of Coq modules to be required prior to the
definition of bindings. This allows for specifying Coq models that are
not loaded by Coq by default.

We give an example of this feature with the [Str] module in the
example folder; this module contains one function, introduced by the
[str.mli] file, whose model is itself defined in the [StrModels]
module.